Constante de Kaprekar terminada, con paquetes y llamadas
